tests/functional/requirements.txt
pytest==7.4.3
pytest-asyncio==0.12.0
aiohttp==3.7.2
elasticsearch==8.10.1
elasticsearch[async]==8.10.1
redis==5.0.1
pydantic==2.4.2
pydantic_settings==2.0.3
python-dotenv==1.0.0